requirements.txt
42cc-pystyle==0.0.15
Django==1.6.7
Pillow==3.0.0
South==1.0.2
django-imagekit==3.2.7
pilkit==1.1.13
six==1.10.0
